+++ js/App.js
@@ -1,6 +1,6 @@
 import CardGame from "./pages/CardGame.js";
 import Home from "./pages/Home.js";
-import { init } from "./utils/router.js";
+import { init } from "./lib/utils/router.js";
 
 export default function App({ $target }) {
   this.route = () => {
@@ -19,5 +19,6 @@ export default function App({ $target }) {
   init(this.route);
 
   this.route();
+
   window.addEventListener("popstate", this.route);
 }
