+++ js/pages/Home.js
@@ -1,4 +1,4 @@
-import { routeChange } from "../utils/router.js";
+import { routeChange } from "../lib/utils/router.js";
 
 export default function Home({ $target }) {
   const $page = document.createElement("div");
